requirements.txt
# Web framework
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.1
gunicorn==22.0.0
python-dotenv==1.0.1
Werkzeug==3.0.3

# ML / NLP
torch==2.3.1
transformers==4.41.2
datasets==2.19.1
accelerate==0.30.1
sentencepiece==0.2.0
vaderSentiment==3.3.2

# Data
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
openpyxl==3.1.4
tqdm==4.66.4

# Voice / Audio
SpeechRecognition==3.10.4
pydub==0.25.1
# PyAudio is required for microphone input — install separately:
# Windows: pip install pyaudio  (or download wheel from https://www.lfd.uci.edu/~gohlke/pythonlibs/#pyaudio)
# Linux:   sudo apt-get install portaudio19-dev && pip install pyaudio
# Mac:     brew install portaudio && pip install pyaudio

# Utilities
psutil==6.0.0
python-dateutil==2.9.0

# Testing
pytest==8.2.2
pytest-flask==1.3.0
